feat: create service to handle sending GA4 events
