+++ gatsby/src/pages/shop/checkout/success.tsx
@@ -11,6 +11,7 @@ import Seo from "../../../components/seo"
 import { StripePrice } from "../../../models/stripe"
 
 import { retrieveCheckoutSession } from "../../../api"
+import { sendGA4PurchaseEvent } from "../../../services/ga4"
 
 import * as styles from "./success.module.scss"
 
@@ -37,32 +38,7 @@ const Success = ({
             if (!session) return
 
             setSession(session)
-
-            const discount =
-                session.total_details?.breakdown?.discounts[0]?.discount
-
-            const gtagEventData: Gtag.EventParams = {
-                transaction_id: sessionId,
-                value: (session.amount_total || 0) / 100,
-                ...(session.currency
-                    ? { currency: session.currency.toUpperCase() }
-                    : {}),
-                ...(discount?.coupon.name
-                    ? { coupon: discount.coupon.name }
-                    : {}),
-                items: session.line_items?.data.map(item => ({
-                    item_id: item.price?.product.metadata.sku,
-                    item_name: item.price?.product.name,
-                    item_category: item.price?.product.metadata.category,
-                    price: item.amount_total / 100,
-                    discount: item.amount_discount / 100,
-                    quantity: item.quantity ?? 1,
-                })),
-            }
-
-            if ("gtag" in window) {
-                window.gtag("event", "purchase", gtagEventData)
-            }
+            sendGA4PurchaseEvent({ session })
         }
 
         fetchSession()
